+++ tests/changelogs/CHANGELOG.md
@@ -2,8 +2,18 @@
 
 ## [Unreleased]
 
+### Removed
+
+- A test removal line
+- And another removal
+
 ## [0.1.0] - 2022-01-13
 
 ### Added
 
 - Created the package
+
+### Changed
+
+- Changed something in the package
+- Updated the target framework